requirements.txt
Django==4.2
djangorestframework==3.14.0
Pillow==9.3.0
PyMuPDF==1.21.1